dalle/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==21.2.0

